+++ src/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteValues.String.cs
@@ -24,34 +24,6 @@ namespace System.Text.Json
         public void WriteStringValue(string value, bool escape = true)
            => WriteStringValue(value.AsSpan(), escape);
 
-        // TODO: Move to separate file
-        public void WriteElementValue(JsonElement value)
-        {
-            ValidateWritingValue();
-            ReadOnlySpan<char> escapedValue = value.GetRawText().AsSpan();
-
-            int idx = 0;
-            WriteListSeparator(ref idx);
-
-            ReadOnlySpan<byte> byteSpan = MemoryMarshal.AsBytes(escapedValue);
-            int partialConsumed = 0;
-            while (true)
-            {
-                OperationStatus status = JsonWriterHelper.ToUtf8(byteSpan.Slice(partialConsumed), _buffer.Slice(idx), out int consumed, out int written);
-                idx += written;
-                if (status == OperationStatus.Done)
-                {
-                    break;
-                }
-                partialConsumed += consumed;
-                AdvanceAndGrow(ref idx);
-            }
-            Advance(idx);
-
-            SetFlagToAddListSeparatorBeforeNextItem();
-            _tokenType = ToTokenType(value.Type);
-        }
-
         /// <summary>
         /// Writes the UTF-16 text value (as a JSON string) as an element of a JSON array.
         /// </summary>
